Create a better structure for client/server code
Server code is compiled to bin and client to a single file
public/client.js. Both build tasks are run at the same time on the build
command (no watching sadly).

Because each src directory has its own tsconfig.json it is required to
reference the typings file with a /// reference in the server file.